Added add template method to class template and inheritance to django class and user class
