+++ src/_user/_user.py
@@ -1,11 +1,6 @@
 import os
 import shutil
+from _template.template import Template
 
-class User:
-    def __init__(self, project_path):
-        self.path = project_path
-
-    def add_template(self, template: str):
-        path = os.path.dirname(os.path.abspath(__file__))
-        path += f"\\{template}"
-        shutil.copytree(path, self.path + "\\user")
+class User(Template):
+    pass
